Applied builder pattern + Pool inside Offer
The classes have constructors with many parameters, and keeping track of their order is ridiculous. Therefore, I decided to use the builder pattern (that we have learned it for something at the university)

The 'Offer' structure has also changed. Some attributes have been included in the 'pool' object, in order to no repeat data and to facilitate the use of Mustache in the HTMLs

I've made so many things the other day and today that trying to separate the commits would require a significant amount of time and effort (usually, I would spend it, but I'm burned out)