Standardized and streamlined pool logic
Applied 'offer' CSS styling to 'pool' display and resolved issues stemming from removing the message functions that were in 'DataBase'.
Modified logic to ensure consistent display of the display comment input line, now always visible, and simplified code structure. However, my intention was to eliminate nMessages and idM. Still, due to the necessary changes in 'Message' and its usage, I opted to defer this task for another day. Additionally, we need better control over how comments are loaded because the current method is too crude.
Definitely, I was trying to do some changes to, finally, leave it as it was, which resulted in wasted time, especially when the line where the codes breaks is the one that you've changed moments before.
It's funny that this is the "final" commit, considering it's one of the changes that I've made the first day but postponed to the next day.