Implemented offers pagination
Resize between loads will have unintended consequences, elements will not load sequentially.
Implementing 'hasMore' and 'nextPage' on the server-side will be nice.

Obs: The use of @RequestMapping is interesting as it provides a header to all routes of the controller, which can be useful for giving consistent semantics to the routes.