Refactored, reorganized and formatted the HTMLs
Introduced 'data' tag in pool HTML and 'form' tag in form HTMLs for CSS usage.
Changed the cancel button of edit user to redirect to the profile.
Added eye icon on proposals and enhanced titles of some containers.
Improved handling of messages for empty offers in the profile; eliminated the need for 'type==x' conditionals, now based on the element 'x' itself.