Implemented Google Maps in Angular
Based on a Google example and stashed since April 30th,  I faced difficulties displaying anything.
Following advice from a professor, I abandoned the use of a separate module.
Over the weekend, I identified and corrected some errors (offers not iterable, restrictions in google cloud...), allowing me to successfully display the map and the URJC marker.
However, I encountered an issue with the geocoder, which is displaying a "REQUEST_DENIED: API keys with referer restrictions cannot be used with this API" error.
The error prevents the geolocation of addresses, so if I manage to pass the latitude and longitude via API REST, I can resolve this issue.