Implemented edit offer
Stashed all the afternoon because there was no 'type' in the builder.
Implemented a lifeguards reset when the offer is updated.
Refactored 'checkOfferRequest' to analyze only the necessary fields when editing.
Implemented 'formatedDate' function to prevent issues with dates; if there's an error, it returns the local date.
Removed unnecessary attributes from the HTML and added necessary ones.
Added redirection to the created offer, which was missing.